Create an array of timestamp matches
